+++ gamepad.h
@@ -43,7 +43,7 @@ namespace io
         uint8_t axis[3]{0x80, 0x80, 0x80};
         Hat hat{Hat::RELEASED};
         uint32_t buttons{0};
-
+        char GamePadName[20]{};
     public:
         void convertButtonsFromAxis(int axisX, int axisY);
         void convertButtonsFromHat();