+++ hid_app.cpp
@@ -5,6 +5,7 @@
 
 #include <tusb.h>
 #include <stdio.h>
+#include "xinput_host.h"
 #include "gamepad.h"
 
 #ifdef __cplusplus
@@ -19,19 +20,35 @@ extern "C"
         uint8_t _report_count[CFG_TUH_HID];
         tuh_hid_report_info_t _report_info_arr[CFG_TUH_HID][MAX_REPORT];
 
+        // Is dual shock 4 controller detected?
         bool isDS4(uint16_t vid, uint16_t pid)
         {
             return vid == 0x054c && (pid == 0x09cc || pid == 0x05c4);
         }
-
+        // Is dual sense controller detected?
         bool isDS5(uint16_t vid, uint16_t pid)
         {
             return vid == 0x054c && pid == 0x0ce6;
         }
-
+        // Is PSClassic controller detected?
+        bool isPSClassic(uint16_t vid, uint16_t pid)
+        {
+            return vid == 0x054c && pid == 0x0cda;
+        }
+        // Is Nintendo controller detected?
         bool isNintendo(uint16_t vid, uint16_t pid)
         {
-            return vid == 0x057e && ( pid == 0x2009 || pid == 0x2017);
+            return vid == 0x057e && (pid == 0x2009 || pid == 0x2017);
+        }
+        // Is Genesis Mini 1 controller or Genesis Mini 2 controller detected?
+        bool isGenesisMini(uint16_t vid, uint16_t pid)
+        {
+            return vid == 0x0ca3 && (pid == 0x0025 || pid == 0x0024);
+        }
+        // Is MantaPad detected? Cheap Aliexpress SNES/NES controller
+        bool isMantaPad(uint16_t vid, uint16_t pid)
+        {
+            return vid == 0x081f && pid == 0xe401;
         }
 
         struct DS4Report
@@ -104,16 +121,146 @@ extern "C"
 
             int getHat() const { return buttons[0] & 15; }
         };
+
+        // Report for Genesis Mini controller
+        struct GenesisMiniReport
+        {
+            uint8_t byte1;
+            uint8_t byte2;
+            uint8_t byte3;
+            uint8_t byte4;
+            uint8_t byte5;
+            uint8_t byte6;
+            uint8_t byte7;
+            uint8_t byte8;
+            struct Button
+            {
+
+                inline static constexpr int A = 0b01000000;
+                inline static constexpr int B = 0b00100000;
+                inline static constexpr int C = 0b00000010;
+                inline static constexpr int START = 0b00100000;
+                inline static constexpr int UP = 0;
+                inline static constexpr int DOWN = 0b11111111;
+                inline static constexpr int LEFT = 0;
+                inline static constexpr int RIGHT = 0b11111111;
+                ;
+            };
+        };
+
+        // Report for MantaPad, cheap AliExpress SNES controller
+        struct MantaPadReport
+        {
+            uint8_t byte1;
+            uint8_t byte2;
+            uint8_t byte3;
+            uint8_t byte4;
+            uint8_t byte5;
+            uint8_t byte6;
+            uint8_t byte7;
+            uint8_t byte8;
+
+            struct Button
+            {
+                inline static constexpr int A = 0b00100000;
+                inline static constexpr int SNESB = 0b01000000;
+                inline static constexpr int X = 0b00010000;
+                inline static constexpr int NESB = X; // B Button on NES controller is X on SNES controller
+                inline static constexpr int Y = 0b10000000;
+                inline static constexpr int SELECT = 0b00010000;
+                inline static constexpr int START = 0b00100000;
+                inline static constexpr int UP = 0b00000000;
+                inline static constexpr int DOWN = 0b11111111;
+                inline static constexpr int LEFT = 0b00000000;
+                inline static constexpr int RIGHT = 0b11111111;
+                inline static constexpr int SHOULDERLEFT = 0b00000001;
+                inline static constexpr int SHOULDERRIGHT = 0b00000010;
+            };
+        };
+        // Report for PSClassic controller
+        struct PSClassicReport
+        {
+            uint8_t buttons;
+            // Idle      00010100
+            // up        00000100
+            // upright   00001000
+            // right     00011000
+            // rightdown 00101000
+            // down      00100100
+            // downleft  00100000
+            // left      00010000
+            // leftup    00100000
+            // start     00010110
+            // select    00010101
+            // St + sel  00010111
+            // selectup  00000101
+            // selectdown 00100101
+            uint8_t hat;
+            struct Button
+            {
+                inline static constexpr int ButtonsIdle = 0x00;
+                inline static constexpr int HatIdle = 0b00010100;
+                inline static constexpr int Circle = 0x02;
+                inline static constexpr int Cross = 0x04;
+                inline static constexpr int SELECT = 0b00010101;
+                inline static constexpr int START = 0b00010110;
+                inline static constexpr int UP = 0b00000100;
+                inline static constexpr int UPRIGHT = 0b00001000;
+                inline static constexpr int RIGHT = 0b00011000;
+                inline static constexpr int RIGHTDOWN = 0b00101000;
+                inline static constexpr int DOWN = 0b00100100;
+                inline static constexpr int DOWNLEFT = 0b00100000;
+                inline static constexpr int LEFT = 0b00010000;
+                inline static constexpr int LEFTUP = 0b00000000;
+                inline static constexpr int SELECTUP = 0b00000101;
+                inline static constexpr int SELECTDOWN = 0b00100101;
+                inline static constexpr int SELECTSTART = 0b00010111;
+            };
+        };
     }
 
     void tuh_hid_mount_cb(uint8_t dev_addr, uint8_t instance, uint8_t const *desc_report, uint16_t desc_len)
     {
         uint16_t vid, pid;
+        auto &gp = io::getCurrentGamePadState(0);
         tuh_vid_pid_get(dev_addr, &vid, &pid);
 
-        printf("HID device address = %d, instance = %d is mounted\n", dev_addr, instance);
+        if (isDS4(vid, pid))
+        {
+            printf("Dual Shock 4 Controller detected - device address = %d, instance = %d is mounted - ", dev_addr, instance);
+            strcpy(gp.GamePadName, "Dual Shock 4");
+        }
+        else if (isDS5(vid, pid))
+        {
+            printf("Dual Sense Controller detected - device address = %d, instance = %d is mounted - ", dev_addr, instance);
+            strcpy(gp.GamePadName, "Dual Sense");
+        }
+        else if (isMantaPad(vid, pid))
+        {
+            printf("MantaPad detected - device address = %d, instance = %d is mounted - ", dev_addr, instance);
+            strcpy(gp.GamePadName, "Manta");
+        }
+        else if (isGenesisMini(vid, pid))
+        {
+            printf("Sega Mega Drive/Genesis Mini %d controller detected - device address = %d, instance = %d is mounted - ", (pid == 0x0025) ? 1 : 2, dev_addr, instance);
+            sprintf(gp.GamePadName, "Genesis Mini %d", (pid == 0x0025) ? 1 : 2);
+        }
+        else if (isPSClassic(vid, pid))
+        {
+            printf("PlayStation Classic controller detected - device address = %d, instance = %d is mounted - ", dev_addr, instance);
+            strcpy(gp.GamePadName, "PSClassic");
+        }
+        // else if (isNintendo(vid, pid))
+        // {
+        //     printf("(Unsupported) Nintendo controller detected - device address = %d, instance = %d is mounted - ", dev_addr, instance);
+        //     strcpy(gp.GamePadName, "Nintendo");
+        // }
+        else
+        {
+            printf("Unkown device detected - HID device address = %d, instance = %d is mounted - ", dev_addr, instance);
+            sprintf(gp.GamePadName, "%04x:%04x", vid, pid);
+        }
         printf("VID = %04x, PID = %04x\r\n", vid, pid);
-
         const char *protocol_str[] = {"None", "Keyboard", "Mouse"}; // hid_protocol_type_t
         uint8_t const interface_protocol = tuh_hid_interface_protocol(dev_addr, instance);
 
@@ -134,6 +281,7 @@ extern "C"
         // Assume the controller is disconnected
         auto &gp = io::getCurrentGamePadState(0);
         gp.flagConnected(false);
+        gp.GamePadName[0] = 0;
     }
 
     void tuh_hid_report_received_cb(uint8_t dev_addr,
@@ -214,9 +362,119 @@ extern "C"
                 return;
             }
         }
+        else if (isMantaPad(vid, pid))
+        {
+            if (sizeof(MantaPadReport) == len)
+            {
+                auto r = reinterpret_cast<const MantaPadReport *>(report);
+                auto &gp = io::getCurrentGamePadState(0);
+                gp.buttons =
+                    (r->byte6 & MantaPadReport::Button::A ? io::GamePadState::Button::A : 0) |
+                    (r->byte6 & MantaPadReport::Button::NESB ? io::GamePadState::Button::B : 0) |
+                    (r->byte6 & MantaPadReport::Button::SNESB ? io::GamePadState::Button::B : 0) |
+                    (r->byte7 & MantaPadReport::Button::START ? io::GamePadState::Button::START : 0) |
+                    (r->byte7 & MantaPadReport::Button::SELECT ? io::GamePadState::Button::SELECT : 0) |
+                    (r->byte2 == MantaPadReport::Button::UP ? io::GamePadState::Button::UP : 0) |
+                    (r->byte2 == MantaPadReport::Button::DOWN ? io::GamePadState::Button::DOWN : 0) |
+                    (r->byte1 == MantaPadReport::Button::LEFT ? io::GamePadState::Button::LEFT : 0) |
+                    (r->byte1 == MantaPadReport::Button::RIGHT ? io::GamePadState::Button::RIGHT : 0);
+                gp.flagConnected(true);
+            }
+            else
+            {
+                printf("Invalid MantaPad report size %zd\n", len);
+                return;
+            }
+        }
+        else if (isGenesisMini(vid, pid))
+        {
+            if (sizeof(GenesisMiniReport) == len)
+            {
+                auto r = reinterpret_cast<const GenesisMiniReport *>(report);
+                auto &gp = io::getCurrentGamePadState(0);
+                // Swap A and B because the button layout is different from NES controller
+                gp.buttons =
+                    (r->byte6 & GenesisMiniReport::Button::B ? io::GamePadState::Button::A : 0) |
+                    (r->byte6 & GenesisMiniReport::Button::A ? io::GamePadState::Button::B : 0) |
+                    (r->byte7 & GenesisMiniReport::Button::C ? io::GamePadState::Button::SELECT : 0) |
+                    (r->byte7 & GenesisMiniReport::Button::START ? io::GamePadState::Button::START : 0) |
+                    (r->byte5 == GenesisMiniReport::Button::UP ? io::GamePadState::Button::UP : 0) |
+                    (r->byte5 == GenesisMiniReport::Button::DOWN ? io::GamePadState::Button::DOWN : 0) |
+                    (r->byte4 == GenesisMiniReport::Button::LEFT ? io::GamePadState::Button::LEFT : 0) |
+                    (r->byte4 == GenesisMiniReport::Button::RIGHT ? io::GamePadState::Button::RIGHT : 0);
+                gp.flagConnected(true);
+            }
+            else
+            {
+                printf("Invalid Genesis Mini report size %zd\n", len);
+                return;
+            }
+        }
+        else if (isPSClassic(vid, pid))
+        {
+            if (sizeof(PSClassicReport) == len)
+            {
+                auto r = reinterpret_cast<const PSClassicReport *>(report);
+                auto &gp = io::getCurrentGamePadState(0);
+                gp.buttons =
+                    (r->buttons & PSClassicReport::Button::Cross ? io::GamePadState::Button::B : 0) |
+                    (r->buttons & PSClassicReport::Button::Circle ? io::GamePadState::Button::A : 0);
+
+                switch (r->hat)
+                {
+                case PSClassicReport::Button::UP:
+                    gp.buttons = gp.buttons | io::GamePadState::Button::UP;
+                    break;
+                case PSClassicReport::Button::UPRIGHT:
+                    gp.buttons = gp.buttons | io::GamePadState::Button::UP | io::GamePadState::Button::RIGHT;
+                    break;
+                case PSClassicReport::Button::RIGHT:
+                    gp.buttons = gp.buttons | io::GamePadState::Button::RIGHT;
+                    break;
+                case PSClassicReport::Button::RIGHTDOWN:
+                    gp.buttons = gp.buttons | io::GamePadState::Button::RIGHT | io::GamePadState::Button::DOWN;
+                    break;
+                case PSClassicReport::Button::DOWN:
+                    gp.buttons = gp.buttons | io::GamePadState::Button::DOWN;
+                    break;
+                case PSClassicReport::Button::DOWNLEFT:
+                    gp.buttons = gp.buttons | io::GamePadState::Button::DOWN | io::GamePadState::Button::LEFT;
+                    break;
+                case PSClassicReport::Button::LEFT:
+                    gp.buttons = gp.buttons | io::GamePadState::Button::LEFT;
+                    break;
+                case PSClassicReport::Button::LEFTUP:
+                    gp.buttons = gp.buttons | io::GamePadState::Button::LEFT | io::GamePadState::Button::UP;
+                    break;
+                case PSClassicReport::Button::SELECT:
+                    gp.buttons = gp.buttons | io::GamePadState::Button::SELECT;
+                    break;
+                case PSClassicReport::Button::START:
+                    gp.buttons = gp.buttons | io::GamePadState::Button::START;
+                    break;
+                case PSClassicReport::Button::SELECTSTART:
+                    gp.buttons = gp.buttons | io::GamePadState::Button::SELECT | io::GamePadState::Button::START;
+                    break;
+                case PSClassicReport::Button::SELECTUP:
+                    gp.buttons = gp.buttons | io::GamePadState::Button::SELECT | io::GamePadState::Button::UP;
+                    break;
+                case PSClassicReport::Button::SELECTDOWN:
+                    gp.buttons = gp.buttons | io::GamePadState::Button::SELECT | io::GamePadState::Button::DOWN;
+                    break;
+                default:
+                    break;
+                }
+                gp.flagConnected(true);
+            }
+            else
+            {
+                printf("Invalid PSClassic Mini report size %zd\n", len);
+                return;
+            }
+        }
         else if (isNintendo(vid, pid))
         {
-            printf("Nintendo: len = %d\n", len);
+            printf("Nintendo: len = %d\n", len); // Nintendo controllers do not report back
         }
         else
         {
@@ -257,10 +515,48 @@ extern "C"
                 switch (rpt_info->usage)
                 {
                 case HID_USAGE_DESKTOP_KEYBOARD:
-                    TU_LOG1("HID receive keyboard report\n");
-                    // Assume keyboard follow boot report layout
-                    //                process_kbd_report((hid_keyboard_report_t const *)report);
+                {
+                    auto r = reinterpret_cast<const hid_keyboard_report_t *>(report);
+                    auto &gp = io::getCurrentGamePadState(0);
+                    strcpy(gp.GamePadName, "Keyboard");
+                    gp.buttons = 0;
+                    for (uint8_t i = 0; i < 6; i++)
+                    {
+                        if (r->keycode[i])
+                        {
+                            switch (r->keycode[i])
+                            {
+                            case HID_KEY_A:
+                                gp.buttons |= io::GamePadState::Button::SELECT;
+                                break;
+                            case HID_KEY_S:
+                                gp.buttons |= io::GamePadState::Button::START;
+                                break;
+                            case HID_KEY_Z:
+                                gp.buttons |= io::GamePadState::Button::B;
+                                break;
+                            case HID_KEY_X:
+                                gp.buttons |= io::GamePadState::Button::A;
+                                break;
+                            case HID_KEY_ARROW_UP:
+                                gp.buttons |= io::GamePadState::Button::UP;
+                                break;
+                            case HID_KEY_ARROW_DOWN:
+                                gp.buttons |= io::GamePadState::Button::DOWN;
+                                break;
+                            case HID_KEY_ARROW_LEFT:
+                                gp.buttons |= io::GamePadState::Button::LEFT;
+                                break;
+                            case HID_KEY_ARROW_RIGHT:
+                                gp.buttons |= io::GamePadState::Button::RIGHT;
+                                break;
+                            default:
+                                break;
+                            }
+                        }
+                    }
                     break;
+                }
 
                 case HID_USAGE_DESKTOP_MOUSE:
                     TU_LOG1("HID receive mouse report\n");
@@ -307,7 +603,118 @@ extern "C"
             printf("Error: cannot request to receive report\r\n");
         }
     }
+#pragma region XINPUT
+    // Since https://github.com/hathach/tinyusb/pull/2222, we can add in custom vendor drivers easily
+    usbh_class_driver_t const *usbh_app_driver_get_cb(uint8_t *driver_count)
+    {
+        *driver_count = 1;
+        return &usbh_xinput_driver;
+    }
 
+    // XXInput type of controlles. (Xbox 360, Xbox One, Xbox Series X)
+    // This is somewhat flaky and might not work with all controllers.
+    // Tested devices:
+    //    - xbox Series X controller : Works
+    //    - xbox One controller : Works
+    //    - xbox elite controller : Works
+    //    - 8bitdo SN30 Pro+ V6.01: Works. Hold X + Start to switch to Xinput mode. (LED 1 and 2 will blink). Then connect to USB.
+    //    - 8bitdo Pro 2 V3.04: Works. Hold X + Start to switch to Xinput mode. (LED 1 and 2 will blink). Then connect to USB.
+    //    - 8bitdo SN30 PRO Wired : Not working, recognized but no report
+    //    - 8bitdo SF30 v2.05 Pro : Works Hold X + Start to switch to Xinput mode. (LED 1 and 2 will blink). Then connect to USB.
+    //    - 8bitdo SN30 v2.05 Pro : Not tested, should probably work
+    //
+    // Troubleshooting:
+    //  After flashing some bigger games, the controller might become unresponsive:
+    //      - XBOX Controller. Play always with batteries removed. When controller becomes unresponsive:
+    //              - unplug and replug the controller.
+    //              - If controller is still unresponsive, unplug the pico from power, wait a few seconds then plug it back in. 
+    //              - Press start on the controller to start the last flashed game.
+    //     - 8bitdo controllers, when controller becomes unresponsive:
+    //              - Disconnect the controller
+    //              - Hold start to switch the controller off (if it has built-in battery).
+    //              - reconnect the controller.
+    //              - Press start to start the last flashed game.
+    void tuh_xinput_report_received_cb(uint8_t dev_addr, uint8_t instance, xinputh_interface_t const *xid_itf, uint16_t len)
+    {
+        const xinput_gamepad_t *p = &xid_itf->pad;
+        const char *type_str;
+
+        if (xid_itf->last_xfer_result == XFER_RESULT_SUCCESS)
+        {
+            switch (xid_itf->type)
+            {
+            case 1:
+                type_str = "Xbox One";
+                break;
+            case 2:
+                type_str = "Xbox 360 Wireless";
+                break;
+            case 3:
+                type_str = "Xbox 360 Wired";
+                break;
+            case 4:
+                type_str = "Xbox OG";
+                break;
+            default:
+                type_str = "Unknown";
+            }
+
+            if (xid_itf->connected && xid_itf->new_pad_data)
+            {
+
+                auto &gp = io::getCurrentGamePadState(0);
+                gp.buttons = 0;
+                if (p->wButtons & XINPUT_GAMEPAD_A)
+                    gp.buttons |= io::GamePadState::Button::B;
+                if (p->wButtons & XINPUT_GAMEPAD_B)
+                    gp.buttons |= io::GamePadState::Button::A;
+
+                if (p->wButtons & XINPUT_GAMEPAD_DPAD_UP)
+                    gp.buttons |= io::GamePadState::Button::UP;
+                if (p->wButtons & XINPUT_GAMEPAD_DPAD_DOWN)
+                    gp.buttons |= io::GamePadState::Button::DOWN;
+                if (p->wButtons & XINPUT_GAMEPAD_DPAD_LEFT)
+                    gp.buttons |= io::GamePadState::Button::LEFT;
+                if (p->wButtons & XINPUT_GAMEPAD_DPAD_RIGHT)
+                    gp.buttons |= io::GamePadState::Button::RIGHT;
+                if (p->wButtons & XINPUT_GAMEPAD_START)
+                    gp.buttons |= io::GamePadState::Button::START;
+                if (p->wButtons & XINPUT_GAMEPAD_BACK)
+                    gp.buttons |= io::GamePadState::Button::SELECT;
+                if (p->wButtons & XINPUT_GAMEPAD_GUIDE)
+                    gp.buttons |= (io::GamePadState::Button::START | io::GamePadState::Button::SELECT);
+                gp.flagConnected(true);
+            }
+        }
+        tuh_xinput_receive_report(dev_addr, instance);
+    }
+
+    void tuh_xinput_mount_cb(uint8_t dev_addr, uint8_t instance, const xinputh_interface_t *xinput_itf)
+    {
+        auto &gp = io::getCurrentGamePadState(0);
+        strcpy(gp.GamePadName, "XInput");
+        printf("XINPUT MOUNTED %02x %d\n", dev_addr, instance);
+        // If this is a Xbox 360 Wireless controller we need to wait for a connection packet
+        // on the in pipe before setting LEDs etc. So just start getting data until a controller is connected.
+        if (xinput_itf->type == XBOX360_WIRELESS && xinput_itf->connected == false)
+        {
+            tuh_xinput_receive_report(dev_addr, instance);
+            return;
+        }
+        tuh_xinput_set_led(dev_addr, instance, 0, true);
+        tuh_xinput_set_led(dev_addr, instance, 1, true);
+        tuh_xinput_set_rumble(dev_addr, instance, 0, 0, true);
+        tuh_xinput_receive_report(dev_addr, instance);
+    }
+
+    void tuh_xinput_umount_cb(uint8_t dev_addr, uint8_t instance)
+    {
+        auto &gp = io::getCurrentGamePadState(0);
+        gp.GamePadName[0] = 0;
+        gp.flagConnected(false);
+        printf("XINPUT UNMOUNTED %02x %d\n", dev_addr, instance);
+    }
+#pragma endregion
 #ifdef __cplusplus
 }
 #endif