+++ tusb_config.h
@@ -29,6 +29,9 @@ extern "C"
 #define CFG_TUSB_HOST_DEVICE_MAX (4 + 1)
 #define CFG_TUH_HID_EP_BUFSIZE 64
 
+// allow to use XInput
+#define CFG_TUH_XINPUT 1
+//#define CFG_TUSB_DEBUG 0
 #ifdef __cplusplus
 }
 #endif